requirements.txt
jinja2>=3.1
pyyaml>=6.0
markdown>=3.5
pillow>=10.0
weasyprint>=60.0
